Change datalist to options list. Renamed Meetings.
The decision to change the datalist to options list was made because it
will offer a simpler programming experience when it comes to programming
the push notifications and the in app reminders. Seeing a though this
app is for a specific type of individual(steward) it makes more since
for the code to be rigid with predefined options instead of unlimited
options with suggestions.