Express Backend Connected with React.
This is this perfect time to commit, while the fix is fresh in my mind.
I couldn't connect express with react because the clause to catch status
errors was incomplete. I simply put response instead of response.status
This prevented the fetch function from checking the status of the
response effectively.