dotenv working
The dotenv issue has been resolved. The issue lies in finding the
correct path to the .env file. It is not always as simple as calling
require('dotenv').config() and expecting it the go to the correct path.
Depending on the level of nesting you may have to go back a few folders
in order to access the dotenv file.