User Info rendering on HomePage
Display name rendering on homepage. Next will pass down to the component
I want to contain it. Also in order to get this functional I had to give
cors Options set the origin to the localhost:3000. Next I had to modify
the fetch request to fetch strict to /userInfo instead of prepending it
with the 5000 port