SignUp form is Posting, Error Handling works,
Sign up for is posting now and I'm able to see the data in the console.
I had to install the bodyparser package from npm in order to see the
information. Javascript client side error handling is working and the
next step is to make the errors visible for the client.