UserTasks and meeting coming from DB
Had to figure out whether it's a best practice to populate with mongodb
or to have a one to many relationship that I can simply fetch from using
the current users _id. In this case one to many is a better practice and
will prevent errors where pushing to userTasks array could loose sync
with data in userTask collection. Why over complicate