Fixed composition errors. Working on List display
Compositions errors were causing the tailwind h-screen class to collide
with other sections of the app. I removed areas where the size of the
components were not designated correctly and extracted code from
components that could easily live inside of other components to prevent
the passing down of variables to multiple components.